+++ build_web_compilers/web/stack_trace_mapper.dart
@@ -81,6 +81,7 @@ class LazyMapping extends Mapping {
 
   LazyMapping(this._provider);
 
+  // ignore: unreachable_from_main
   List toJson() => _bundle.toJson();
 
   @override